Implement DrawingArea::set_draw_func() manually to allow FnMut closures instead of just Fn
Fixes https://github.com/gtk-rs/gtk4-rs/issues/548
