+++ gtk4/Gir.toml
@@ -944,10 +944,7 @@ status = "generate"
 manual_traits = ["DrawingAreaExtManual"]
     [[object.function]]
     name = "set_draw_func"
-        [[object.function.parameter]]
-        name = "draw_func"
-        # we add manual unset_draw_func
-        nullable = false
+    manual = true
 
 [[object]]
 name = "Gtk.DropDown"
