+++ gtk4/src/auto/drawing_area.rs
@@ -381,9 +381,6 @@ pub trait DrawingAreaExt: 'static {
     #[doc(alias = "gtk_drawing_area_set_content_width")]
     fn set_content_width(&self, width: i32);
 
-    #[doc(alias = "gtk_drawing_area_set_draw_func")]
-    fn set_draw_func<P: Fn(&DrawingArea, &cairo::Context, i32, i32) + 'static>(&self, draw_func: P);
-
     #[doc(alias = "resize")]
     fn connect_resize<F: Fn(&Self, i32, i32) + 'static>(&self, f: F) -> SignalHandlerId;
 
@@ -412,45 +409,6 @@ impl<O: IsA<DrawingArea>> DrawingAreaExt for O {
     fn set_content_width(&self, width: i32) {
         unsafe {
             ffi::gtk_drawing_area_set_content_width(self.as_ref().to_glib_none().0, width);
-        }
-    }
-
-    fn set_draw_func<P: Fn(&DrawingArea, &cairo::Context, i32, i32) + 'static>(
-        &self,
-        draw_func: P,
-    ) {
-        let draw_func_data: Box_<P> = Box_::new(draw_func);
-        unsafe extern "C" fn draw_func_func<
-            P: Fn(&DrawingArea, &cairo::Context, i32, i32) + 'static,
-        >(
-            drawing_area: *mut ffi::GtkDrawingArea,
-            cr: *mut cairo::ffi::cairo_t,
-            width: libc::c_int,
-            height: libc::c_int,
-            user_data: glib::ffi::gpointer,
-        ) {
-            let drawing_area = from_glib_borrow(drawing_area);
-            let cr = from_glib_borrow(cr);
-            let callback: &P = &*(user_data as *mut _);
-            (*callback)(&drawing_area, &cr, width, height);
-        }
-        let draw_func = Some(draw_func_func::<P> as _);
-        unsafe extern "C" fn destroy_func<
-            P: Fn(&DrawingArea, &cairo::Context, i32, i32) + 'static,
-        >(
-            data: glib::ffi::gpointer,
-        ) {
-            let _callback: Box_<P> = Box_::from_raw(data as *mut _);
-        }
-        let destroy_call3 = Some(destroy_func::<P> as _);
-        let super_callback0: Box_<P> = draw_func_data;
-        unsafe {
-            ffi::gtk_drawing_area_set_draw_func(
-                self.as_ref().to_glib_none().0,
-                draw_func,
-                Box_::into_raw(super_callback0) as *mut _,
-                destroy_call3,
-            );
         }
     }
